requirements.txt
Pillow==9.2.0
altair==4.2.0
clarifai==9.8.1
numpy==1.23.1
pandas==1.4.3
requests==2.28.1
stqdm==0.0.4
streamlit==1.22.0
vega-datasets==0.9.0
st-annotated-text==4.0.1
